requirements.txt
boto3>=1.34.84
python-dotenv>=1.0.1
requests==2.31.0
langchain-google-genai>=1.0.3
langchain-tavily
langgraph
langchain